+++ src/commands/init_config.rs
@@ -87,6 +87,7 @@ pub const DEFAULT_JSON: &str = r#"{
     "biome": { "enabled": true, "ignore-files": [] },
     "checkstyle": { "enabled": true },
     "delete_empty_folders": { "enabled": true },
+    "dprint": { "enabled": false, "ignore-files": [] },
     "gherkin_lint": { "enabled": true, "ignore-files": [] },
     "ghokin": { "enabled": true, "ignore-files": [] },
     "git_diff_check": { "enabled": true },
@@ -254,6 +255,11 @@ mod tests {
                     }),
                     delete_empty_folders: Some(ApplicationNoFile {
                         enabled: Some(true),
+                        operations: None,
+                    }),
+                    dprint: Some(ApplicationWithFile {
+                        enabled: Some(false),
+                        ignore_files: Some(vec![]),
                         operations: None,
                     }),
                     gherkin_lint: Some(ApplicationWithFile {