+++ src/stacks/jsonc.rs
@@ -1,3 +1,4 @@
+use crate::apps::dprint::Dprint;
 use crate::apps::prettier::Prettier;
 use crate::domain::{Fix, Lint, Stack, StackType};
 use std::path::Path;
@@ -18,7 +19,7 @@ impl Stack for JsonC {
     }
 
     fn fixes(&self) -> Vec<Box<dyn Fix>> {
-        vec![Box::new(Prettier {})]
+        vec![Box::new(Dprint {}), Box::new(Prettier {})]
     }
 }