+++ src/stacks/yml.rs
@@ -1,3 +1,4 @@
+use crate::apps::dprint::Dprint;
 use crate::apps::{actionlint::Actionlint, prettier::Prettier};
 use crate::domain::{Fix, Lint, Stack, StackType};
 use std::path::Path;
@@ -19,7 +20,7 @@ impl Stack for Yml {
     }
 
     fn fixes(&self) -> Vec<Box<dyn Fix>> {
-        vec![Box::new(Prettier {})]
+        vec![Box::new(Dprint {}), Box::new(Prettier {})]
     }
 }